Add build for @nteract/types
