+++ packages/types/redux.js
@@ -1,8 +1,1 @@
-// @flow
-/* eslint-disable no-undef */
-
-declare type Action = {
-  +type: string
-};
-
-declare type Dispatch<A: Action> = (action: A) => A;
+module.exports = require("./redux");
