+++ packages/types/src/channels.js
@@ -1,0 +1,3 @@
+// @flow
+
+declare type Channels = rxjs$Subject<*>;
