+++ packages/types/src/redux.js
@@ -1,0 +1,8 @@
+// @flow
+/* eslint-disable no-undef */
+
+declare type Action = {
+  +type: string
+};
+
+declare type Dispatch<A: Action> = (action: A) => A;
